Fuck

+++ src/main/resources/assets/nvidium/shaders/occlusion/section_raster/fragment.glsl
@@ -12,12 +12,17 @@ layout(early_fragment_tests) in;
 #ifdef DEBUG
 layout(location = 0) out vec4 colour;
 void main() {
-    uint uid = bitfieldReverse(gl_PrimitiveID*132471+123571);
-    colour = vec4(float((uid>>0)&7)/7, float((uid>>3)&7)/7, float((uid>>6)&7)/7, 1.0);
-    sectionVisibility[gl_PrimitiveID>>8] = uint8_t(gl_PrimitiveID);
+    uint uid = bitfieldReverse(uint(gl_PrimitiveID) * 132471u + 123571u);
+    colour = vec4(
+    float((uid >> 0u) & 7u) / 7.0,
+    float((uid >> 3u) & 7u) / 7.0,
+    float((uid >> 6u) & 7u) / 7.0,
+    1.0
+    );
+    sectionVisibility[uint(gl_PrimitiveID) >> 8u] = uint8_t(uint(gl_PrimitiveID));
 }
 #else
 void main() {
-    sectionVisibility[gl_PrimitiveID>>8] = uint8_t(gl_PrimitiveID);
+    sectionVisibility[uint(gl_PrimitiveID) >> 8u] = uint8_t(uint(gl_PrimitiveID));
 }
 #endif
